requirements.txt
praw==7.2.0
python-telegram-bot==13.3
